+++ src/js/components/charge/coinCharge.js
@@ -1,14 +1,12 @@
 import { $ } from '../../utils/querySelector.js';
-import { COIN_UNITS, STANDARD, STORAGE_NAME } from '../../utils/constants.js';
+import { AMOUNT_ID, COIN_UNITS, STORAGE_NAME } from '../../utils/constants.js';
 import { isValidInputAmount } from '../../utils/validation.js';
-import { showConvertedCoins, coinChargeTemplate } from './coinChargeTemplate.js';
+import { showConvertedCoins } from '../../view/coinCharge.js';
 import { getLocalStorage, setLocalStorage } from '../../utils/storage.js';
 import { showCurrentAmount } from '../../view/view.js';
+import { calculationCurrentAmount } from '../../utils/calculation.js';
 
-let convertedCoins = { 500: 0, 100: 0, 50: 0, 10: 0 };
-const vendingMachineChargeAmountId = '#vending-machine-charge-amount';
-
-const convertAmountIntoCoins = (amount) => {
+const convertAmountIntoCoins = (convertedCoins, amount) => {
   let remainAmount = amount;
   while (remainAmount > 0) {
     // eslint-disable-next-line no-loop-func
@@ -20,17 +18,7 @@ const convertAmountIntoCoins = (amount) => {
   }
 };
 
-const calculationCurrentAmount = (storedChargeCoins) => {
-  let currentAmount = STANDARD.CURRENT_MONEY;
-  // eslint-disable-next-line no-restricted-syntax
-  for (const unit in storedChargeCoins) {
-    currentAmount += storedChargeCoins[unit] * unit;
-  }
-  return currentAmount;
-};
-
-const handleCoinChargeSubmit = (event) => {
-  event.preventDefault();
+export const handleCoinChargeSubmit = (convertedCoins) => {
   let chargedCoin = Number($('#vending-machine-charge-input').value);
   const storedChargeCoins = getLocalStorage(STORAGE_NAME.COIN);
 
@@ -38,23 +26,9 @@ const handleCoinChargeSubmit = (event) => {
     return;
   }
 
-  convertAmountIntoCoins(chargedCoin);
+  convertAmountIntoCoins(convertedCoins, chargedCoin);
   chargedCoin += calculationCurrentAmount(storedChargeCoins);
-  showCurrentAmount(vendingMachineChargeAmountId, chargedCoin);
+  showCurrentAmount(AMOUNT_ID.MACHINE, chargedCoin);
   showConvertedCoins(convertedCoins);
   setLocalStorage(STORAGE_NAME.COIN, convertedCoins);
-};
-
-export const showManageMenu = () => {
-  $('#app-container').innerHTML = coinChargeTemplate;
-  const storedChargeCoins = getLocalStorage(STORAGE_NAME.COIN);
-
-  if (storedChargeCoins) {
-    convertedCoins = storedChargeCoins;
-    showConvertedCoins(storedChargeCoins);
-    const currentAmount = calculationCurrentAmount(storedChargeCoins);
-    showCurrentAmount(vendingMachineChargeAmountId, currentAmount);
-  }
-
-  $('form').addEventListener('submit', handleCoinChargeSubmit);
 };
